vmx: round up timer value on shift

+++ include/cpulocal.hpp
@@ -110,7 +110,7 @@ struct alignas(PAGE_SIZE) Per_cpu {
     mword vmcs_fix_cr4_clr;
     mword vmcs_fix_cr4_mon;
 
-    uint64 vmx_timer_shift;
+    uint8 vmx_timer_shift;
 
     // SVM-related variables
     Paddr    vmcb_root;
